Debug epsilon issue with Matrix.
There was some floating point error resulting from row addition that was
tricking the Matrix into thinking there was a leading one in a column
that did not actually have a leading one, so it improperly calculated
the RREF as a result. Introduced an epsilon equality function and used
it in both rref and equals.